+++ ui/src/data_explorer/containers/DataExplorer.tsx
@@ -32,6 +32,7 @@ import defaultQueryConfig from 'src/utils/defaultQueryConfig'
 
 import {Source, QueryConfig, TimeRange} from 'src/types'
 import {ErrorHandling} from 'src/shared/decorators/errors'
+import LoadData from 'src/loudml/components/LoadData';
 
 interface Props {
   source: Source
@@ -191,12 +192,16 @@ export class DataExplorer extends PureComponent<Props, State> {
       autoRefresh,
       onManualRefresh,
       handleChooseAutoRefresh,
+      source,
     } = this.props
 
     return (
       <>
         <GraphTips />
         <OneClickML />
+        <LoadData
+          source={source}
+          />
         <div
           className="btn btn-sm btn-default"
           onClick={this.handleOpenWriteData}