+++ ui/src/loudml/apis/index.js
@@ -191,3 +191,12 @@ export const getTemplates = async () => {
         throw error
     }
 }
+
+export const nab = datasource => {
+    return AJAX({
+        method: 'POST',
+        url: `/loudml/api/_nab`,
+        data: {datasource},
+        excludeBasepath: true,
+    })
+}