+++ ui/src/loudml/components/Feature.js
@@ -22,10 +22,6 @@ import {
 
 import {
     DEFAULT_METRICS,
-    DEFAULT_IO,
-    DEFAULT_SCORES,
-    DEFAULT_TRANSFORM,
-    DEFAULT_LOUDML_RP
 } from 'src/loudml/constants'
 import {DEFAULT_ANOMALY_TYPE} from 'src/loudml/constants/anomaly'
 
@@ -41,7 +37,6 @@ const safeValue = (collection, value) => {
 
 const EnabledWatermarkValue = ({
     value,
-    // name,
     onEdit
 }) => {
     return (
@@ -57,8 +52,6 @@ const EnabledWatermarkValue = ({
 }
 
 const WatermarkComponent = ({
-    // feature,
-    // name,
     watermark,
     onEdit,
     disabled,
@@ -74,7 +67,6 @@ const WatermarkComponent = ({
     return (
         <EnabledWatermarkValue
             value={watermark}
-            // name={name}
             onEdit={onEdit}
             />
     )
@@ -125,6 +117,7 @@ class Feature extends Component {
             database,
             feature: {measurement},
             measurements,
+            retentionPolicy,
         } = this.props
 
         if (
@@ -136,7 +129,7 @@ class Feature extends Component {
         }
 
         try {
-            const {data} = await showFieldKeys(source.links.proxy, database, measurement, DEFAULT_LOUDML_RP)
+            const {data} = await showFieldKeys(source.links.proxy, database, measurement, retentionPolicy)
             const {errors, fieldSets} = showFieldKeysParser(data)
             if (errors.length) {
                 console.error('Error parsing fields keys: ', errors)
@@ -241,6 +234,7 @@ class Feature extends Component {
             onConfirm,
             feature,
             database,
+            retentionPolicy,
             source,
             locked,
         } = this.props
@@ -282,7 +276,7 @@ class Feature extends Component {
                                         tags={feature.match_all}
                                         database={database}
                                         measurement={feature.measurement}
-                                        retentionPolicy={DEFAULT_LOUDML_RP}
+                                        retentionPolicy={retentionPolicy}
                                         source={source}
                                         onChooseTag={this.onChooseTag}
                                         disabled={locked}
@@ -300,7 +294,6 @@ class Feature extends Component {
         const {
             feature,
             measurements,
-            timeseries,
             locked,
         } = this.props
 
@@ -308,7 +301,7 @@ class Feature extends Component {
             fields,
         } = this.state
 
-        const base = [
+        return [
             {
                 label: 'Measurement',
                 customClass: "col-xs-5",
@@ -362,6 +355,24 @@ class Feature extends Component {
                 />)
             },
             {
+                label: 'Low watermark',
+                customClass: "col-md-5",
+                component: (<WatermarkComponent
+                    watermark={feature.low_watermark}
+                    onEdit={this.handleWatermarkValue('low_watermark')}
+                    disabled={locked}
+                    />)
+            },
+            {
+                label: 'High watermark',
+                customClass: "col-xs-offset-1 col-xs-5",
+                component: (<WatermarkComponent
+                    watermark={feature.high_watermark}
+                    onEdit={this.handleWatermarkValue('high_watermark')}
+                    disabled={locked}
+                    />)
+            },
+            {
                 label: 'Anomaly type',
                 customClass: "col-xs-4",
                 component: (<FeatureDropdown
@@ -374,77 +385,7 @@ class Feature extends Component {
                     disabled={locked}
                 />)
             },
-            {
-                label: 'Scores',
-                customClass: "col-xs-4",
-                component: (<FeatureDropdown
-                    name="scores"
-                    onChoose={this.handleValueChoose('scores')}
-                    items={DEFAULT_SCORES}
-                    selected={safeValue(DEFAULT_SCORES, feature.scores)}
-                    className="dropdown-stretch"
-                    buttonSize="btn-sm"
-                    disabled={locked}
-                />)
-            },
-            {
-                label: 'Transform',
-                customClass: "col-xs-3",
-                component: (<FeatureDropdown
-                    name="transform"
-                    onChoose={this.handleValueChoose('transform')}
-                    items={DEFAULT_TRANSFORM}
-                    selected={safeValue(DEFAULT_TRANSFORM, feature.transform)}
-                    className="dropdown-stretch"
-                    buttonSize="btn-sm"
-                    disabled={locked}
-                />)
-            }
         ]
-
-        if (timeseries) {
-            return [
-                ...base,
-                {
-                    label: 'Input/Output',
-                    customClass: "col-xs-4",
-                    component: (<FeatureDropdown
-                        name="io"
-                        onChoose={this.handleValueChoose('io')}
-                        items={DEFAULT_IO}
-                        selected={safeValue(DEFAULT_IO, feature.io)}
-                        className="dropdown-stretch"
-                        buttonSize="btn-sm"
-                        disabled={locked}
-                    />)
-                }
-            ]
-        }
-
-        return [
-            ...base,
-            {
-                label: 'Low watermark',
-                customClass: "col-md-5",
-                component: (<WatermarkComponent
-                    // feature={feature}
-                    watermark={feature.low_watermark}
-                    onEdit={this.handleWatermarkValue('low_watermark')}
-                    disabled={locked}
-                    />)
-            },
-            {
-                label: 'High watermark',
-                customClass: "col-xs-offset-1 col-xs-5",
-                component: (<WatermarkComponent
-                    // feature={feature}
-                    watermark={feature.high_watermark}
-                    onEdit={this.handleWatermarkValue('high_watermark')}
-                    disabled={locked}
-                    />)
-            }
-        ]
-
     }
 }
 
@@ -452,12 +393,10 @@ const {func, shape, arrayOf, string, bool, number, oneOfType} = PropTypes
 
 EnabledWatermarkValue.propTypes = {
     value: number,
-    // name: string.isRequired,
     onEdit: func.isRequired,
 }
 
 WatermarkComponent.propTypes = {
-    // feature: shape({}).isRequired,
     watermark: number,
     onEdit: func.isRequired,
     disabled: bool.isRequired,
@@ -470,7 +409,6 @@ FeatureDropdown.propTypes = {
 
 Feature.propTypes = {
     feature: shape({}).isRequired,
-    timeseries: bool.isRequired,
     onDelete: func.isRequired,
     onEdit: func.isRequired,
     onKeyDown: func.isRequired,
@@ -478,6 +416,7 @@ Feature.propTypes = {
     measurements: arrayOf(string),
     source: shape(),
     database: string,
+    retentionPolicy: string.isRequired,
     locked: bool.isRequired,
 }