+++ ui/src/loudml/utils/datasource.js
@@ -1,10 +1,29 @@
-export const findSource = (sources, datasource) => {
-    // remove datasource protocol
-    const loudmlHosts = datasource.addr.split('://')
-    const loudmlHost = loudmlHosts.slice(-1)[0]
+const splitAddr = (addr, port) => {
+    // extract host and port from url address
+    const re = /(https?:)?(\/\/)?([\w\.]*):?(\d*)?/
+    const res = re.exec(addr)
+    return {
+        host: res[3],
+        port: res[4]||port,
+    }
+}
+
+const equalAddr = (a, b) => ((a.host===b.host)&&(a.port===b.port))
+
+export const findSource = (sources, datasource, port) => {
+    const loudml = splitAddr(datasource.addr, port)
 
     return sources.find(s => {
-        const sourceDbs = s.url.split('://')
-        return sourceDbs.slice(-1)[0] === loudmlHost
+        const sourceDb = splitAddr(s.url, port)
+        return equalAddr(sourceDb, loudml)
+    })
+}
+
+export const findDatabases = (source, datasources, port) => {
+    const sourceDb = splitAddr(source.url, port)
+
+    return datasources.filter(d => {
+        const loudml = splitAddr(d.addr, port)
+        return equalAddr(sourceDb, loudml)
     })
 }