+++ ui/src/loudml/utils/query.js
@@ -5,7 +5,6 @@ import FeaturesUtils from 'src/loudml/components/FeaturesUtils'
 import { DEFAULT_LOUDML_RP } from 'src/loudml/constants';
 
 const QUERY_CONFIG = {
-    retentionPolicy: DEFAULT_LOUDML_RP,
     areTagsAccepted: true,
     rawText: null,
     range: {
@@ -63,16 +62,23 @@ const getTags = feature => {
     }, {})
 }
 
-const createErrorQueryConfig = (prefix, model, database) => {
-    const {name, features} = model
+const getErrorTags = (feature, name) => {
+    return Object.assign(
+        getTags(feature),
+        {model: [name]})
+}
+
+const createErrorQueryConfig = (prefix, model, datasource) => {
+    const {features, name} = model
     const feature = features[0]
 
     return {
         ...QUERY_CONFIG,
         fields: createErrorQueryFields(prefix, model),
-        database,
-        measurement: `prediction_${name}`,
-        tags: getTags(feature),
+        database: datasource.database,
+        retentionPolicy: datasource.retention_policy||DEFAULT_LOUDML_RP,
+        measurement: 'loudml',
+        tags: getErrorTags(feature, name),
         fill: null,
         groupBy: {
             time: model.bucket_interval,
@@ -81,7 +87,7 @@ const createErrorQueryConfig = (prefix, model, database) => {
     }
 }
 
-const createModelQueryConfig = (model, database) => {
+const createModelQueryConfig = (model, datasource) => {
     const {features} = model
     const feature = features[0]
     const measurement = feature.measurement
@@ -89,7 +95,8 @@ const createModelQueryConfig = (model, database) => {
     return {
         ...QUERY_CONFIG,
         fields: createModelQueryFields(model),
-        database,
+        database: datasource.database,
+        retentionPolicy: datasource.retention_policy||DEFAULT_LOUDML_RP,
         measurement,
         tags: getTags(feature),
         fill: feature.default,
@@ -100,7 +107,7 @@ const createModelQueryConfig = (model, database) => {
     }
 }
 
-export const createQueryFromModel = (model, source, database) => {
+export const createQueryFromModel = (model, source, datasource) => {
     const {settings} = model
     const {links: {self}} = source
     
@@ -116,16 +123,16 @@ export const createQueryFromModel = (model, source, database) => {
         { queryConfig: createErrorQueryConfig(
             'lower',
             m,
-            database),
+            datasource),
         },
         { queryConfig: createModelQueryConfig(
             m,
-            database),
+            datasource),
         },
         { queryConfig: createErrorQueryConfig(
             'upper',
             m,
-            database),
+            datasource),
         },
     ]
     return [{