+++ ui/src/loudml/CHANGELOG.md
@@ -7,8 +7,14 @@
 1.  [#23](https://github.com/regel/chronograf/pull/23): Scores and Transform settings in Feature panel
 1.  [#24](https://github.com/regel/chronograf/pull/24): Datasink
 1.  [#28](https://github.com/regel/chronograf/pull/28): Unserialized Features in model
+1.  [#35](https://github.com/regel/chronograf/pull/35): normalize span for 1-Click ML
 1.  [#36](https://github.com/regel/chronograf/pull/36): Model builder page (Use template)
 1.  [#38](https://github.com/regel/chronograf/pull/38): tags (key/values) in 1-Click model name
+1.  [#39](https://github.com/regel/chronograf/pull/39): retention policy value from Loud ML datasource
+1.  [#41](https://github.com/regel/chronograf/pull/41): Donut model type
+1.  [#42](https://github.com/regel/chronograf/pull/42): Donut settings
+1.  [#43](https://github.com/regel/chronograf/pull/43): About message for OSS
+1.  [#44](https://github.com/regel/chronograf/pull/44): Load nab dataset
 
 ### UI Improvements
 
@@ -19,6 +25,7 @@
 1.  [#28](https://github.com/regel/chronograf/pull/28): Model name input/display in Model Page
 1.  [#28](https://github.com/regel/chronograf/pull/28): Icons for Play/Stop
 1.  [#32](https://github.com/regel/chronograf/pull/32): Error graph picto
+1.  [#40](https://github.com/regel/chronograf/pull/40): 1-Click ML becomes Create baseline
 
 ### Bug Fixes